backend/requirements.txt
paho-mqtt>=2.1.0        # MQTT 통신 라이브러리 (v2.0 이상 필수)
pymysql>=1.1.1          # MySQL 연결 (순수 파이썬 구현체)
cryptography>=43.0.0    # MySQL 8.0/9.0 보안 인증 플러그인 대응 (필수)